+++ src/main/java/io/github/mortuusars/chalk/data/ModBlockTagsProvider.java
@@ -1,0 +1,14 @@
+package io.github.mortuusars.chalk.data;
+
+import io.github.mortuusars.chalk.Chalk;
+import net.minecraft.data.BlockTagsProvider;
+import net.minecraft.data.DataGenerator;
+import net.minecraftforge.common.data.ExistingFileHelper;
+
+import javax.annotation.Nullable;
+
+public class ModBlockTagsProvider extends BlockTagsProvider {
+    public ModBlockTagsProvider(DataGenerator generator, @Nullable ExistingFileHelper existingFileHelper) {
+        super(generator, Chalk.MOD_ID, existingFileHelper);
+    }
+}