+++ src/main/java/io/github/mortuusars/chalk/Chalk.java
@@ -11,12 +11,10 @@ import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-// The value here should match an entry in the META-INF/mods.toml file
 @Mod("chalk")
 public class Chalk
 {
     public static final String MOD_ID = "chalk";
-
     public static final Logger LOGGER = LogManager.getLogger();
 
     public Chalk() {