+++ src/main/java/io/github/mortuusars/chalk/config/CommonConfig.java
@@ -12,15 +12,19 @@ public class CommonConfig {
 
     public static final ForgeConfigSpec.IntValue CHALK_DURABILITY;
     public static final ForgeConfigSpec.ConfigValue<List<? extends String>> GLOWING_ITEMS;
+    public static final ForgeConfigSpec.BooleanValue GENERATE_IN_CHESTS;
 
     public static final List<String> GLOWING_ITEMS_DEFAULT = Arrays.asList("minecraft:glowstone_dust", "cavesandcliffs:glow_ink_sac", "upgrade_aquatic:glowing_ink_sac");
 
     static {
-        CHALK_DURABILITY = BUILDER.comment("How many marks you can draw with single chalk:")
+        CHALK_DURABILITY = BUILDER.comment("How many marks you can draw with single chalk. Default: 64")
                                   .defineInRange("ChalkUses", 64, 0, Integer.MAX_VALUE);
 
         GLOWING_ITEMS = BUILDER.comment("List of items that can make mark glow. \"modid:itemRegistryName\"\nDefault values: " + GLOWING_ITEMS_DEFAULT)
                                .define("GlowItems", GLOWING_ITEMS_DEFAULT, cfgList -> validateGlowItems(cfgList));
+
+        GENERATE_IN_CHESTS = BUILDER.comment("If enabled Chalks will generate in some chests: \nDungeon, Abandoned Mineshaft, Planes and Savanna villages, Cartographer house\nDefault: true")
+                                    .define("ShouldGenerateInChests", true);
 
         SPEC = BUILDER.build();
     }