+++ src/main/java/io/github/mortuusars/chalk/loot/LootTableModification.java
@@ -1,6 +1,7 @@
 package io.github.mortuusars.chalk.loot;
 
 import io.github.mortuusars.chalk.Chalk;
+import io.github.mortuusars.chalk.config.CommonConfig;
 import net.minecraft.loot.LootPool;
 import net.minecraft.loot.TableLootEntry;
 import net.minecraft.util.ResourceLocation;
@@ -10,6 +11,10 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 public class LootTableModification {
     @SubscribeEvent
     public static void LootTablesLoad(final LootTableLoadEvent event){
+
+        if (!CommonConfig.GENERATE_IN_CHESTS.get())
+            return;
+
         if (event.getName().toString().equals("minecraft:chests/abandoned_mineshaft") ||
             event.getName().toString().equals("minecraft:chests/simple_dungeon")){