+++ CHANGELOG.md
@@ -4,9 +4,10 @@ Added ❌-mark: drawn by Sneak-RightClicking.
 - Enhanced Pick Block (Middle Click) - it will now select proper chalk from your inventory 
 - Glowing Inc Sac from Upgrade Aquatic can be used to make mark glowing
 - Added dye tags to all chalks. You can now draw on Supplementaries Blackboard with them. Enable colored_blackboard in Supplementaries config for better results. 
-- Added all chalks to "forge:chalk" tag.
+- Added all chalks to "forge:chalk" tag
 - Better offhand interactions: if you have something in offhand while drawing the mark - offhand item will not be used
-- Marks are now colored programmatically. 
+- Marks are now colored programmatically
+- Toned down Glowing mark particles frequency 
 
 ---